transcendence/requirements.txt
Django>=3.0,<4.0
djangorestframework
djangorestframework-jwt
drf-yasg
channels 
python-dotenv>=0.19.0
flask_security 
psycopg2-binary>=2.8
Pillow==10.2.0